% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beer.R
\name{beer}
\alias{beer}
\title{An internal function for geom_beer}
\usage{
beer(x, y, color = "#bdbdbd", lineend = "butt")
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{color}{indicates the color. The default color is set to #bdbdbd which is grrey-ish}

\item{lineend}{Line end style (round, butt, square). Default is set to butt.}
}
\value{
mapping: x = ~-x + 1.5, y = ~y, xend = ~x - 1.5, yend = ~y
geom_segment: arrow = NULL, arrow.fill = NULL, lineend = butt, linejoin = round, na.rm = FALSE
stat_identity: na.rm = FALSE
position_identity
}
\description{
An internal function for geom_beer
}
