% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bun.R
\name{bun}
\alias{bun}
\title{An internal function for geom_hotdog}
\usage{
bun(x, y, color = "#CF8F41")
}
\arguments{
\item{x}{consumes numeric vector}

\item{y}{consumes numeric vector}

\item{color}{indicates the color. The default color is set to #CF8F41 which is beig-ish}
}
\value{
mapping: x = ~-x + 1.5, y = ~y, xend = ~x - 1.5, yend = ~y
geom_segment: arrow = NULL, arrow.fill = NULL, lineend = butt, linejoin = round, na.rm = FALSE
stat_identity: na.rm = FALSE
position_identity #'
}
\description{
An internal function for geom_hotdog
}
\examples{
bun(4)
}
