% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_beer.R
\name{geom_beer}
\alias{geom_beer}
\title{geom_beer}
\usage{
geom_beer(length = 10)
}
\arguments{
\item{length}{Consumes numeric vector}
}
\value{
Returns an image of a beer can when used with ggplot2
}
\description{
geom_beer
}
