% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_gin.R
\name{geom_gin}
\alias{geom_gin}
\title{Create a gin bottle usinf geom_gin and ggplot2}
\usage{
geom_gin(length = 10)
}
\arguments{
\item{length}{Consumes numeric vector}
}
\value{
Returns an image of a beer can when used with ggplot2
}
\description{
Create a gin bottle usinf geom_gin and ggplot2
}
