% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hotdog.R
\name{geom_hotdog}
\alias{geom_hotdog}
\title{You can create a yummy hotdog in ggplot2 with geom_hotdog}
\usage{
geom_hotdog(length = 10)
}
\arguments{
\item{length}{Consumes numeric vector}
}
\value{
An image of a hot dog when used together with ggplot
}
\description{
You can create a yummy hotdog in ggplot2 with geom_hotdog
}
\examples{
geom_hotdog(5)
}
