% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_redwine.R
\name{geom_redwine}
\alias{geom_redwine}
\title{Creates an image of a red wine bottle}
\usage{
geom_redwine(length = 10)
}
\arguments{
\item{length}{Consumes numeric vector}
}
\value{
An image of a boottle of red wine when used with ggplot2
}
\description{
Creates an image of a red wine bottle
}
\examples{
geom_redwine(5)
}
