% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_whitewine.R
\name{geom_whitewine}
\alias{geom_whitewine}
\title{Creates an image of a red wine bottle}
\usage{
geom_whitewine(length = 10)
}
\arguments{
\item{length}{Consumes numeric vector}
}
\value{
Returns an image of a bottle of white wine can when used with ggplot2
}
\description{
Creates an image of a red wine bottle
}
\examples{
geom_whitewine(10)
}
