% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wine.R
\name{wine}
\alias{wine}
\title{An internal function for geom_redwine and geom_whitewine and geom_gin}
\usage{
wine(x, y, color = "#00441b", lineend = "butt")
}
\arguments{
\item{x}{consumes a numeric vector}

\item{y}{consumes a numeric vector}

\item{color}{indicates the color. The default color is set to #00441b which is burgandy-ish}

\item{lineend}{Line end style (round, butt, square). Default is set to butt.}
}
\value{
mapping: x = ~-x + 1.5, y = ~y, xend = ~x - 1.5, yend = ~y
geom_segment: arrow = NULL, arrow.fill = NULL, lineend = butt, linejoin = round, na.rm = FALSE
stat_identity: na.rm = FALSE
position_identity
}
\description{
An internal function for geom_redwine and geom_whitewine and geom_gin
}
\examples{
wine(9)
}
